Hooked up initial out-of-the-box task management API in various mono-repo services: - udp2rediscluster - MsmqMetrics - FilePusher - ProcessMetrics - redisinfo - UdpViaTcp - pubsub2pubsub
Initial implementation per this commit merely exposes some basic task management related information, however down the line this will allow for the data driven task defintions

+++ plugins/VFS/vfspluginAWSS3/include/vfspluginAWSS3_CS3Archive.h
@@ -118,7 +118,7 @@ class GUCEF_HIDDEN CS3Archive : public VFS::CArchive
     virtual bool GetFileMetaData( const VFS::CString& filePath      ,
                                   CORE::CResourceMetaData& metaData ) const GUCEF_VIRTUAL_OVERRIDE;
 
-    virtual VFS::UInt32 GetFileSize( const VFS::CString& filePath ) const GUCEF_VIRTUAL_OVERRIDE;
+    virtual VFS::UInt64 GetFileSize( const VFS::CString& filePath ) const GUCEF_VIRTUAL_OVERRIDE;
 
     virtual VFS::CString GetFileHash( const VFS::CString& file ) const GUCEF_VIRTUAL_OVERRIDE;
